TECHNICIAN: reschedule screen working

+++ lib/data/model/technician/technician_info_model.dart
@@ -15,8 +15,8 @@ class TechnicianInfo {
   final String? createdDate;
   final String? updatedBy;
   final String? updatedDate;
-  final int? billingAmount;
-  final int? paidAmount;
+  final double? billingAmount;
+  final double? paidAmount;
   final String? technicianCategoryName;
   final String? technicianName;
   final String? propertyName;
